+++ src/DisPlay.Umbraco.EmbeddedContent.Web.UI/embeddedcontentproperty.directive.js
@@ -3,28 +3,31 @@
 
 function EmbeddedContentPropertyDirective(fileManager) {
   return {
-    require: '^ngModel',
+    restrict: 'E',
+    template: '<umb-property property="property" ng-if="ready"><umb-property-editor model="property"></umb-property-editor></umb-property>',
     scope: {
       setFiles: '&',
+      property: '=',
       embeddedContentItem: '='
     },
     link(scope) {
 
-      let alias = scope.model.alias;
-      let id = scope.model.id;
+      let alias = scope.property.alias;
+      let id = scope.property.id;
 
       // return the original alias and id when the property is serialized
-      scope.model.toJSON = function() {
+      scope.property.toJSON = function() {
         return _.extend({}, this, { alias: alias, id: id });
       };
 
       // make sure the property alias and id is unique
-      scope.model.alias = scope.model.id = `item-${scope.embeddedContentItem.key}-${alias}`;
+      scope.property.alias = scope.property.id = `item-${scope.embeddedContentItem.key}-${alias}`;
 
+      scope.ready = true;
       scope.$on('filesSelected', (e, args) => {
-        let filesArray = [].slice.call(args.files).map(file => _.extend(file, { propertyAlias: scope.model.alias }));
-        scope.model.selectedFiles =filesArray.map(file => file.name);
-        fileManager.setFiles(scope.model.alias, []);
+        let filesArray = [].slice.call(args.files).map(file => _.extend(file, { propertyAlias: scope.property.alias }));
+        scope.property.selectedFiles = filesArray.map(file => file.name);
+        fileManager.setFiles(scope.property.alias, []);
         scope.setFiles({ files: filesArray });
       });
     }